refactor: domain / infrastructure 패키지 분리 및 레포지토리 인터페이스 추출
