+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -3,8 +3,8 @@ package kitchenpos.application;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.dao.OrderTableDao;
 import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.OrderTableDao;
 import kitchenpos.domain.table.OrderTableRepository;
 import kitchenpos.domain.table.TableGroup;
 import kitchenpos.domain.table.TableGroupRepository;
