+++ src/main/java/kitchenpos/domain/menu/ProductRepository.java
@@ -1,28 +1,12 @@
 package kitchenpos.domain.menu;
 
 import java.util.List;
-import kitchenpos.dao.ProductDao;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public class ProductRepository {
+public interface ProductRepository {
 
-    private final ProductDao productDao;
+    Product get(Long id);
 
-    public ProductRepository(final ProductDao productDao) {
-        this.productDao = productDao;
-    }
+    Product add(Product product);
 
-    public Product get(final Long id) {
-        return productDao.findById(id)
-                .orElseThrow(IllegalArgumentException::new);
-    }
-
-    public Product add(final Product product) {
-        return productDao.save(product);
-    }
-
-    public List<Product> getAll() {
-        return productDao.findAll();
-    }
+    List<Product> getAll();
 }
