+++ src/test/java/kitchenpos/infrastructure/BaseFakeDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
