+++ src/test/java/kitchenpos/infrastructure/MenuGroupFakeDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure;
 
 import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menu.MenuGroupDao;
 
 public class MenuGroupFakeDao extends BaseFakeDao<MenuGroup> implements MenuGroupDao {
