+++ src/test/java/kitchenpos/infrastructure/ProductFakeDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure;
 
 import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.menu.ProductDao;
 
 public class ProductFakeDao extends BaseFakeDao<Product> implements ProductDao {
 }
