+++ src/test/java/kitchenpos/infrastructure/TableGroupFakeDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure;
 
 import kitchenpos.domain.table.TableGroup;
+import kitchenpos.domain.table.TableGroupDao;
 
 public class TableGroupFakeDao extends BaseFakeDao<TableGroup> implements TableGroupDao {
 }
