+++ src/test/java/kitchenpos/ui/TableGroupRestControllerTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
+import kitchenpos.domain.table.OrderTableDao;
+import kitchenpos.domain.table.TableGroupDao;
 import kitchenpos.fixture.OrderTableFactory;
 import kitchenpos.ui.dto.TableGroupRequest;
 import kitchenpos.ui.dto.TableGroupRequest.TableGroupInnerOrderTable;
