+++ src/test/java/kitchenpos/ui/TableRestControllerTest.java
@@ -4,7 +4,7 @@ import static kitchenpos.fixture.OrderTableFactory.notEmptyTable;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import kitchenpos.dao.OrderTableDao;
+import kitchenpos.domain.table.OrderTableDao;
 import kitchenpos.ui.dto.OrderTableRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
